+++ music_assistant/helpers/web.py
@@ -2,10 +2,8 @@
 
 import asyncio
 import inspect
-import ipaddress
 import re
 from dataclasses import dataclass
-from functools import wraps
 from typing import Any, Callable, Dict, Optional, Tuple, Union, get_args, get_origin
 
 import ujson
